refactor(moka): 对齐 Moka 部门接口契约精简同步链路
- 全链路移除 operatorEmail 字段与参数（Request/SyncRequest/Converter/Client/
  Manager/Controller），push-to-moka 接口不再接收操作人邮箱
- sequence 由 BigDecimal 改为 Integer，校验注解 DecimalMin/DecimalMax 调整为 Min/Max
- 同步更新 Controller/Client 接口 Javadoc 与日志输出

+++ cacch-integration-integration/src/main/java/com/cacch/integration/integration/moka/client/dto/MokaDeptItem.java
@@ -3,7 +3,6 @@ package com.cacch.integration.integration.moka.client.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -59,9 +58,9 @@ public class MokaDeptItem {
     private Integer type;
 
     /**
-     * 部门排序，0~10000 支持两位小数
+     * 部门排序，0~10000
      */
-    private BigDecimal sequence;
+    private Integer sequence;
 
     /**
      * 部门状态：1 正常 / 0 停用
